lab-1337/pyproject.toml
[project]
name = "lab-1337"
version = "0.1.0"
description = "1337 Experiments Lab - Rigorous experiments for the agentic era"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "claude-agent-sdk>=0.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "datasets>=3.0.0",  # HumanEval benchmark
    "pythea @ file:///Users/yza.vyas/oss/pythea",  # Core hallucination detection
    "strawberry @ file:///Users/yza.vyas/oss/pythea/strawberry",  # Trace budget verification
    # REP-002: Mandates vs Motivations
    "swebench>=2.0.0",  # SWE-bench evaluation harness
    "arize-phoenix>=4.0.0",  # Local trace visualization
    "openinference-instrumentation-anthropic>=0.1.0",  # Auto-instrument Claude calls
]

[project.scripts]
lab = "lab.cli:main"        # Primary entry point (new)
lab-1337 = "lab.cli:main"   # Backward compatibility

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/lab"]

